[refactor/#224] modalStore.ts 파일명 Zustand 네이밍 컨벤션 적용
