+++ src/components/Modal/ReservationInfoModal/ReservationInfoModal.tsx
@@ -5,7 +5,7 @@ import {
   useGetMyReservations,
   useGetMyReservedSchedule,
 } from "@/src/hooks/useMyActivities";
-import useModalStore from "@/src/stores/modalStore";
+import useModalStore from "@/src/stores/useModalStore";
 import { ReservationStatusType } from "@/src/types/reservation";
 import { formatDateToKorean } from "@/src/utils/calendarFormatDate";
 import clsx from "clsx";
