+++ src/components/Modal/ReviewModal/ReviewModal.tsx
@@ -1,7 +1,7 @@
 import CloseIcon from "@/assets/svgs/btnXbig.svg";
 import ReviewContent from "@/src/components/Modal/ReviewModal/Common/ReviewContent";
 import ReviewForm from "@/src/components/Modal/ReviewModal/Common/ReviewForm";
-import useModalStore from "@/src/stores/modalStore";
+import useModalStore from "@/src/stores/useModalStore";
 import { ReviewData } from "@/src/types/myReservationsResponses";
 import { formatDate3 } from "@/src/utils/calendarFormatDate";
 import clsx from "clsx";
